refactor(client): invert shared/features layering violations (#1780)
useEventSource (core chat SSE hook) hardcoded the Office add-in's
token key and unconditionally imported officeAuth's refresh logic,
silently coupling every chat stream to the Office feature's token
lifecycle. It now accepts optional getAuthHeaders/onUnauthorized
callbacks; useAppChat forwards them and useOfficeChatAdapter supplies
the Office-specific implementations, while the plain SPA path keeps
its previous authToken-only behavior as the default.

DynamicLanguageEditor and its 5 admin form editors now import
formValidationContext from its new shared/contexts/ location.
