+++ client/src/features/admin/components/GroupFormEditor.jsx
@@ -9,7 +9,7 @@ import {
   isFieldRequired
 } from '../../../utils/schemaValidation';
 import AdminFormErrorSummary from './AdminFormErrorSummary';
-import { FormValidationProvider } from './formValidationContext';
+import { FormValidationProvider } from '../../../shared/contexts/formValidationContext';
 
 /**
  * GroupFormEditor - Form-based editor for group configuration
